feat(spike): widen difftest register bridge to XLEN (64-bit)
RV64 exposed two latent defects in the spike difftest path:
- misa mismatch: the ISA table carried RV32's MXL encoding
- register truncation: the bridge (difftest_regs_t, init, get_*_ptr,
  sync_regs_to_spike) was a fixed 32-bit protocol; an MMIO-step sync
  truncated 64-bit GPRs to their low word, so spike executed mulhu on a
  mangled operand and diverged

C side:
- difftest_regs_t pc/gpr -> uint64_t (protocol width is fixed; the C ABI
  cannot be generic over XLEN), init_pc/init_gpr/get_*_ptr widened
- sync_regs_to_spike keeps the rv32 sign-extension branch (spike stores
  rv32 reg_t sign-extended)

Rust side:
- ffi.rs signatures match the 64-bit ABI
- regs_diff compares masked to XLEN (rv32 sign-extension from spike)
- removed dead DIFFTEST_MAGIC/VERSION macros
- README rewritten to the actual (function-style, spike-owned memory)
  protocol instead of the stale zero-copy design

Verified: microbench test passes under --difftest spike for rv64im and
rv32im, and --difftest remu for both; unit tests 10/10.

+++ remu_simulator/simulators/spike/src/ffi.rs
@@ -6,11 +6,11 @@ use std::os::raw::{c_char, c_int, c_uint};
 
 use libloading::Library;
 
-/// Layout matches difftest_regs_t
+/// Layout matches difftest_regs_t (XLEN-wide values; RV32 keeps low 32 bits)
 #[repr(C)]
 pub(crate) struct DifftestRegs {
-    pub(crate) pc: u32,
-    pub(crate) gpr: [u32; 32],
+    pub(crate) pc: u64,
+    pub(crate) gpr: [u64; 32],
 }
 
 /// Memory layout: base + size only; Spike owns the memory
@@ -32,8 +32,8 @@ pub(crate) struct SpikeFns {
     pub init: unsafe extern "C" fn(
         layout: *const DifftestMemLayout,
         n_regions: usize,
-        init_pc: u32,
-        init_gpr: *const u32,
+        init_pc: u64,
+        init_gpr: *const u64,
         xlen: c_uint,
         isa: *const c_char,
     ) -> SpikeDifftestCtx,
@@ -54,9 +54,9 @@ pub(crate) struct SpikeFns {
     /// Returns 0 success, 1 program exit, -1 error
     pub step: unsafe extern "C" fn(ctx: SpikeDifftestCtx) -> c_int,
 
-    pub get_pc_ptr: unsafe extern "C" fn(ctx: SpikeDifftestCtx) -> *const u32,
+    pub get_pc_ptr: unsafe extern "C" fn(ctx: SpikeDifftestCtx) -> *const u64,
 
-    pub get_gpr_ptr: unsafe extern "C" fn(ctx: SpikeDifftestCtx) -> *const u32,
+    pub get_gpr_ptr: unsafe extern "C" fn(ctx: SpikeDifftestCtx) -> *const u64,
 
     pub get_csr: unsafe extern "C" fn(ctx: SpikeDifftestCtx, csr_addr: u16) -> u32,
